+++ NEWS.md
@@ -210,6 +210,11 @@ Breaking changes
 
 This section lists changes that do not have deprecation warnings.
 
+  * `replace(s::AbstractString, pat=>repl)` for function `repl` arguments formerly
+    passed a substring to `repl` in all cases.  It now passes substrings for
+    string patterns `pat`, but a `Char` for character patterns (when `pat` is a
+    `Char`, collection of `Char`, or a character predicate) ([#25815]).
+
   * `readuntil` now does *not* include the delimiter in its result, matching the
     behavior of `readline`. Pass `keep=true` to get the old behavior ([#25633]).